CliffordAlgebras.CliffordAlgebra — Type

CliffordAlgebra

CliffordAlgebra{Np,Nn,Nz,S} is a type that describes a geometric algebra with the signature (Np,Nn,Nz), base symbols S.

CliffordAlgebras.order — Function

order(::CliffordAlgebra)
order(::Type{<:CliffordAlgebra})

Returns the order of the algebra. The order is the sum of the signature.

CliffordAlgebras.signature — Function

signature(::CliffordAlgebra)
signature(::Type{<:CliffordAlgebra})

Returns the signature of the algebra.

CliffordAlgebras.dimension — Function

dimension(::CliffordAlgebra)
dimension(::Type{<:CliffordAlgebra})

Returns the dimension of the algebra, i.e. the number of coefficients in a general multivector.

CliffordAlgebras.character — Function

character(::CliffordAlgebra)
character(::Type{<:CliffordAlgebra})

Returns the square of the pseudoscalar of the algebra.

CliffordAlgebras.basesymbol — Function

basesymbol(::CliffordAlgebra, n::Integer)
basesymbol(::Type{<:CliffordAlgebra}, n::Integer)

Returns the symbol used for the n-th basis multivector of the algebra.

CliffordAlgebras.MultiVector — Type

MultiVector{CA,T,BI}

Type for a multivector belonging to the algebra CA<:CliffordAlgbra with vector coefficients of type T. Coefficients are stored using a sparse coding, and only the coefficients of the basis indices stored in the tuple BI are considered.

Base.:* — Function

a * b

Calculates the geometric product of two MultiVectors a and b.

Base.:/ — Function

a / b
(/)(a::MultiVector{CA}, b::MuliVector{CA}) where CA

Calculates the MultiVector quotient a/b by evaluating a*inv(b).

Base.:\ — Function

b \ a
(\)(b::MultiVector{CA}, b::MuliVector{CA}) where CA

Calculates the MultiVector quotient a/b by evaluating inv(b)*a.

CliffordAlgebras.:∧ — Function

a ∧ b

Calculates the wedge product between two MultiVectors a and b.

CliffordAlgebras.:∨ — Function

a ∨ b

Calculates the vee product of the MultiVectors a and b.

CliffordAlgebras.:⋅ — Function

a ⋅ b

Calculates the "fat dot" product between the MultiVectors a and b.

CliffordAlgebras.:⨼ — Function

a ⨼ b

Calculates the left contraction of the MultiVectors a and b.

CliffordAlgebras.:⨽ — Function

a ⨽ b

Calculates the right contraction of the MultiVectors a and b.

CliffordAlgebras.:⋆ — Function

a ⋆ b

Calculates the scalar product of the MultiVectors a and b.

CliffordAlgebras.:×₊ — Function

a ×₊ b

Calculates the anti-commutator ab+ba of two MultiVectors a and b.

CliffordAlgebras.:×₋ — Function

a ×₋ b

Calculates the commutator ab-ba of two MultiVectors a and b.

CliffordAlgebras.:≀ — Function

a ≀ b

Calculates the sandwich product abreverse(a) for two MultiVectors a and b.

Base.inv — Function

inv(::MultiVector)

Finds the inverse of the MultiVector. If no inverse exists a SingularException is thrown.

CliffordAlgebras.polarize — Function

polarize(mv::MultiVector)
mv'

Calculates the polarization of the MultiVector, i.e. mv * pseudoscalar.

LinearAlgebra.norm — Function

norm(::MultiVector)

Calculates the MultiVector norm defined as sqrt(grade(mv*reverse(mv),0))

LinearAlgebra.norm_sqr — Function

norm_sqr(::MultiVector)

Calculates the MultiVector squared norm defined as grade(mv*reverse(mv),0)

Base.exp — Function

exp(::MultiVector)

Calculates the exponential function of the MultiVector defined by analytic continuation. The generated code is automaticall specialized for the sparse representation of the MultiVector. It may take advantage of commuting base vectors and split off exponential factors. Hyperbolic, trigonometric and nilpotent solutions are recognized and handled separately. Calling prune or grade before exp may help to find the best algorithm for the exponential evaluation.

CliffordAlgebras.algebra — Function

algebra(::MultiVector)
algebra(::Type{<:MultiVector})

Returns the CliffordAlgebra to which the MultiVector belongs.

CliffordAlgebras.baseindices — Function

baseindices(::MultiVector)
baseindices(::Type{<:MultiVector})

Returns the indices for the sparse MultiVector basis.

CliffordAlgebras.coefficients — Function

coefficients(::MultiVector)

Returns the sparse coefficients of the MultiVector.

CliffordAlgebras.scalar — Function

scalar(mv::MultiVector)

Returns the scalar component of the multivector. The result if of the internal storage type eltype(mv).

CliffordAlgebras.grade — Function

grade(::MultiVector, k::Integer)

Projects the MultiVector onto the k-vectors.

CliffordAlgebras.Λᵏ — Function

Λᵏ(::MultiVector, ::Val{k}) where k
Λᵏ(::MultiVector, k::Integer)

Projects the MultiVector onto k-vectors. Similar to grade(mv,k), but uses @generated code and compile time optimizations.

Base.reverse — Function

reverse(::MultiVector)

Returns the MultiVector that has all the basis vector products reversed.

Base.conj — Function

conj(mv::MultiVector)

Return the conjugate of the MultiVector, i.e. reverse(grin(mv)).

CliffordAlgebras.grin — Function

grin(mv::MultiVector)

Returns the grade involution of the MultiVector, i.e. even(mv) - odd(mv).

CliffordAlgebras.dual — Function

dual(mv::MultiVector)

Returns the Poincaré dual of the MultiVector, such that for all basis MultiVectors mv * dual(mv) = pseudoscalar. Dual is a linear map and the images of other MultiVectors follow from the images of the basis MultiVectors.

CliffordAlgebras.even — Function

even(::MultiVector)

Returns the even grade projection of the MultiVector.

CliffordAlgebras.odd — Function

odd(::MultiVector)

Returns the odd grade projection of the MultiVector.

CliffordAlgebras.basevector — Function

basevector(::CliffordAlgebra, n::Integer)
basevector(::Type{<:CliffordAlgebra}, n::Integer)

Returns the n-th basis MultiVector of the given CliffordAlgebra.

basevector(::CliffordAlgebra, name::Symbol)
basevector(::Type{<:CliffordAlgebra}, name::Symbol)

Returns the basis MultiVector with the specified name from the given Clifford Algebra.

CliffordAlgebras.pseudoscalar — Function

pseudoscalar(::CliffordAlgebra)
pseudoscalar(::Type{<:CliffordAlgebra})

Returns the pseudoscalar of the given algebra.

CliffordAlgebras.prune — Function

prune(::MultiVector ; rtol = 1e-8 )

Returns a new MultiVector with all basis vectors removed from the sparse basis whose coefficients fall below the relative magnitude threshold. This function is not type stable, because the return type depends on the sparse basis.

CliffordAlgebras.extend — Function

extend(::MultiVector)

Returns a new MultiVector with a non-sparse coefficient coding. This can be useful to manage type stability.

CliffordAlgebras.vector — Function

vector(::MultiVector)

Returns the non-sparse vector representation of the MutliVector.

CliffordAlgebras.matrix — Function

matrix(::MultiVector)

Returns the matrix algebra representation of the MultiVector.