Exported symbols from CliffordAlgebras.jl:

Base.:* — Method

a * b

Calculates the geometric product of two MultiVectors a and b.

Base.:/ — Method

a / b
(/)(a::MultiVector{CA}, b::MuliVector{CA}) where CA

Calculates the MultiVector quotient a/b by evaluating a*inv(b).

Base.:\ — Method

b \ a
(\)(b::MultiVector{CA}, b::MuliVector{CA}) where CA

Calculates the MultiVector quotient a/b by evaluating inv(b)*a.

Base.:~ — Method

~a
(~)(::MultiVector)

Returns the reversed MultiVector reverse(a).

Base.conj — Method

conj(mv::MultiVector)

Return the conjugate of the MultiVector, i.e. reverse(grin(mv)).

Base.exp — Method

exp(::MultiVector)

Calculates the exponential function of the MultiVector defined by analytic continuation. The generated code is automaticall specialized for the sparse representation of the MultiVector. It may take advantage of commuting base vectors and split off exponential factors. Hyperbolic, trigonometric and nilpotent solutions are recognized and handled separately. Calling prune or grade before exp may help to find the best algorithm for the exponential evaluation.

Base.inv — Method

inv(::MultiVector)

Finds the inverse of the MultiVector. If no inverse exists a SingularException is thrown.

Base.reverse — Method

reverse(::MultiVector)

Returns the MultiVector that has all the basis vector products reversed.

CliffordAlgebras.:×₊ — Method

a ×₊ b

Calculates the anti-commutator ab+ba of two MultiVectors a and b.

CliffordAlgebras.:×₋ — Method

a ×₋ b

Calculates the commutator ab-ba of two MultiVectors a and b.

CliffordAlgebras.:∧ — Method

a ∧ b

Calculates the wedge product between two MultiVectors a and b.

CliffordAlgebras.:∨ — Method

a ∨ b

Calculates the vee product of the MultiVectors a and b.

CliffordAlgebras.:≀ — Method

a ≀ b

Calculates the sandwich product abreverse(a) for two MultiVectors a and b.

CliffordAlgebras.:⋅ — Method

a ⋅ b

Calculates the "fat dot" product between the MultiVectors a and b.

CliffordAlgebras.:⋆ — Method

a ⋆ b

Calculates the scalar product of the MultiVectors a and b.

CliffordAlgebras.:⨼ — Method

a ⨼ b

Calculates the left contraction of the MultiVectors a and b.

CliffordAlgebras.:⨽ — Method

a ⨽ b

Calculates the right contraction of the MultiVectors a and b.

CliffordAlgebras.algebra — Method

algebra(::MultiVector)
algebra(::Type{<:MultiVector})

Returns the CliffordAlgebra to which the MultiVector belongs.

CliffordAlgebras.baseindices — Method

baseindices(::MultiVector)
baseindices(::Type{<:MultiVector})

Returns the indices for the sparse MultiVector basis.

CliffordAlgebras.basesymbol — Method

basesymbol(::CliffordAlgebra, n::Integer)
basesymbol(::Type{<:CliffordAlgebra}, n::Integer)

Returns the symbol used for the n-th basis multivector of the algebra.

CliffordAlgebras.basevector — Method

basevector(::CliffordAlgebra, n::Integer)
basevector(::Type{<:CliffordAlgebra}, n::Integer)

Returns the n-th basis MultiVector of the given CliffordAlgebra.

CliffordAlgebras.basevector — Method

basevector(::CliffordAlgebra, name::Symbol)
basevector(::Type{<:CliffordAlgebra}, name::Symbol)

Returns the basis MultiVector with the specified name from the given Clifford Algebra.

CliffordAlgebras.character — Method

character(::CliffordAlgebra)
character(::Type{<:CliffordAlgebra})

Returns the square of the pseudoscalar of the algebra.

CliffordAlgebras.coefficients — Method

coefficients(::MultiVector)

Returns the sparse coefficients of the MultiVector.

CliffordAlgebras.dimension — Method

dimension(::CliffordAlgebra)
dimension(::Type{<:CliffordAlgebra})

Returns the dimension of the algebra, i.e. the number of coefficients in a general multivector.

CliffordAlgebras.dual — Method

dual(mv::MultiVector)

Returns the Poincaré dual of the MultiVector, such that for all basis MultiVectors mv * dual(mv) = pseudoscalar. Dual is a linear map and the images of other MultiVectors follow from the images of the basis MultiVectors.

CliffordAlgebras.even — Method

even(::MultiVector)

Returns the even grade projection of the MultiVector.

CliffordAlgebras.extend — Method

extend(::MultiVector)

Returns a new MultiVector with a non-sparse coefficient coding. This can be useful to manage type stability.

CliffordAlgebras.grade — Method

grade(::MultiVector, k::Integer)

Projects the MultiVector onto the k-vectors.

CliffordAlgebras.grin — Method

grin(mv::MultiVector)

Returns the grade involution of the MultiVector, i.e. even(mv) - odd(mv).

CliffordAlgebras.matrix — Method

matrix(::MultiVector)

Returns the matrix algebra representation of the MultiVector.

CliffordAlgebras.odd — Method

odd(::MultiVector)

Returns the odd grade projection of the MultiVector.

CliffordAlgebras.order — Method

order(::CliffordAlgebra)
order(::Type{<:CliffordAlgebra})

Returns the order of the algebra. The order is the sum of the signature.

CliffordAlgebras.polarize — Method

polarize(mv::MultiVector)
mv'

Calculates the polarization of the MultiVector, i.e. mv * pseudoscalar.

CliffordAlgebras.prune — Method

prune(::MultiVector ; rtol = 1e-8 )

Returns a new MultiVector with all basis vectors removed from the sparse basis whose coefficients fall below the relative magnitude threshold. This function is not type stable, because the return type depends on the sparse basis.

CliffordAlgebras.pseudoscalar — Method

pseudoscalar(::CliffordAlgebra)
pseudoscalar(::Type{<:CliffordAlgebra})

Returns the pseudoscalar of the given algebra.

CliffordAlgebras.scalar — Method

scalar(mv::MultiVector)

Returns the scalar component of the multivector. The result if of the internal storage type eltype(mv).

CliffordAlgebras.signature — Method

signature(::CliffordAlgebra)
signature(::Type{<:CliffordAlgebra})

Returns the signature of the algebra.

CliffordAlgebras.vector — Method

vector(::MultiVector)

Returns the non-sparse vector representation of the MutliVector.

CliffordAlgebras.Λᵏ — Method

Λᵏ(::MultiVector, ::Val{k}) where k
Λᵏ(::MultiVector, k::Integer)

Projects the MultiVector onto k-vectors. Similar to grade(mv,k), but uses @generated code and compile time optimizations.

LinearAlgebra.norm — Method

norm(::MultiVector)

Calculates the MultiVector norm defined as sqrt(grade(mv*reverse(mv),0))

LinearAlgebra.norm_sqr — Method

norm_sqr(::MultiVector)

Calculates the MultiVector squared norm defined as grade(mv*reverse(mv),0)

CliffordAlgebras.CliffordAlgebra — Type

CliffordAlgebra

CliffordAlgebra{Np,Nn,Nz,S} is a type that describes a geometric algebra with the signature (Np,Nn,Nz), base symbols S.

CliffordAlgebras.CliffordAlgebra — Method

CliffordAlgebra(Npos::Integer, Nneg::Integer, Nzero::Integer)

Generates a geometric algebra with signature (Npos,Nneg,Nzero).

CliffordAlgebras.CliffordAlgebra — Method

CliffordAlgebra(Npos::Integer, Nneg::Integer)

Generates a geometric algebra with signature (Npos,Nneg,0).

CliffordAlgebras.CliffordAlgebra — Method

CliffordAlgebra(N::Integer)

Generates a geometric algebra with signature (N,0,0).

CliffordAlgebras.CliffordAlgebra — Method

CliffordAlgebra(a::Symbol)

Generates a predefined algebra from a identifier. Known algebras are - :Hyperbolic or :Hyper - :Complex or :ℂ - :Dual or :Grassmann - :Quaternions or :ℍ - :Cl2 and :Cl3 - :Spacetime - :PGA2D or :Projective2D - :PGA3D or :Projective3D - :CGA2D or :Conformal2D - :CGA3D or :Conformal3D - :DCGA3D or :DoubleConformal3D - :TCGA3D or :TripleConformal3D - :DCGSTA or :DoubleConformalSpacetime - :QCGA or :QuadricConformal

CliffordAlgebras.MultiVector — Type

MultiVector{CA,T,BI}

Type for a multivector belonging to the algebra CA<:CliffordAlgbra with vector coefficients of type T. Coefficients are stored using a sparse coding, and only the coefficients of the basis indices stored in the tuple BI are considered.

CliffordAlgebras.MultiVector — Method

MultiVector(::CliffordAlgebra, v::NTuple{N,T}) where {N,T<:Real}
MultiVector(::Type{<:CliffordAlgebra}, v::NTuple{N,T}) where {N,T<:Real}

Creates a MultiVector by converting the provided vector v to a 1-vector. The internal storage type of the MultiVector is T.

CliffordAlgebras.MultiVector — Method

MultiVector(::CliffordAlgebra, a::Real)
MultiVector(::Type{<:CliffordAlgebra}, a::Real)

Creates a MultiVector from the real number a with only a scalar component. The internal storage type of the MultiVector is the type of a.

Private symbols:

CliffordAlgebras.adaptbasefordual — Method

adaptbasefordual(base)

Shuffles the base vector products so that the Poincaré dual of every basis vector is another basis vector.

CliffordAlgebras.basegrade — Method

basegrade(::CliffordAlgebra, n::Integer)
basegrade(::Type{<:CliffordAlgebra}, n::Integer)

Returns the grade of the n-th basis multivector of the algebra.

CliffordAlgebras.basereverse — Method

basereverse(::CliffordAlgebra, n::Integer)
basereverse(::Type{<:CliffordAlgebra}, n::Integer)

Returns the sign change of the n-th basis multivector under reversal.

CliffordAlgebras.basesignature — Method

basesignature(::CliffordAlgebra, n::Integer)
basesignature(::Type{<:CliffordAlgebra}, n::Integer)

Returns the signature value of the n-th basis 1-vector. The return value is +1, -1 or 0.

CliffordAlgebras.basesymbols — Method

basesymbols(::CliffordAlgebra)
basesymbols(::Type{<:CliffordAlgebra})

Returns the 1-vector space basis symbols of the algebra.

CliffordAlgebras.basetable — Method

basetable(::CliffordAlgebra)
basetable(::Type{<:CliffordAlgebra})

Returns the internal basis table of the algebra.

CliffordAlgebras.enumeratebase — Method

enumeratebase(N::Integer)

Generates the basis vector product combinations for the Clifford algebra of an N dimensional vector space.

CliffordAlgebras.findpermutation — Method

findpermutation(a,b)

Find the permutation that takes the a to b.

CliffordAlgebras.multiplicationtable — Method

multiplicationstable(Npos::Integer, Nneg::Integer, Nzero::Integer, base::Tuple)

Generates a multiplication table for the Clifford algebra with the specified signature and base vectors.

CliffordAlgebras.multtable — Method

multtable(::CliffordAlgebra)
multtable(::Type{<:CliffordAlgebra})

Returns the internal multuplication table of the geometric product of the algebra.

CliffordAlgebras.nextselector — Method

nextselector(s::NTuple{N,<:Integer},M) where N

Helper function for enumerating base vectors.